Move tests for okhttp implementation
