+++ open-rosa/src/test/java/org/odk/collect/openrosa/http/CaseInsensitiveEmptyHeadersTest.java
@@ -1,9 +1,7 @@
-package org.odk.collect.android.openrosa;
+package org.odk.collect.openrosa.http;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.odk.collect.openrosa.http.CaseInsensitiveEmptyHeaders;
-import org.odk.collect.openrosa.http.CaseInsensitiveHeaders;
 
 public class CaseInsensitiveEmptyHeadersTest {
     private final CaseInsensitiveHeaders headers = new CaseInsensitiveEmptyHeaders();
