+++ open-rosa/src/test/java/org/odk/collect/openrosa/http/support/MockWebServerHelper.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.openrosa.support;
+package org.odk.collect.openrosa.http.support;
 
 import okhttp3.Request;
 import okhttp3.mockwebserver.MockWebServer;
