+++ open-rosa/src/test/java/org/odk/collect/openrosa/http/support/MockWebServerRule.kt
@@ -1,10 +1,9 @@
-package org.odk.collect.android.rules
+package org.odk.collect.openrosa.http.support
 
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.rules.TestRule
 import org.junit.runner.Description
 import org.junit.runners.model.Statement
-import java.util.ArrayList
 
 class MockWebServerRule : TestRule {
